backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.34.0

# Loguru - 日志
loguru==0.7.3

# 数据验证
pydantic==2.10.0
pydantic-settings==2.7.0

# HTTP 客户端
httpx==0.28.0

# 工具
python-multipart==0.0.20
python-json-logger==3.2.0

# AgentCore 依赖
kuzu==0.8.0
openai==1.60.0
python-dotenv==1.0.0

# 开发依赖
pytest==8.3.0
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.9.0
mypy==1.14.0

